requirements.txt
h5py==2.10.0
keras==2.3.1
librosa
matplotlib
numpy
pyfoal
pypar
pyworld
scipy
soundfile
tensorflow-gpu==1.15
torch
torchcrepe
tqdm
